Java compiler lint warnings fixes in test code
Fixes:
[cast] redundant cast
[rawtypes] found raw type
[serial] serializable class has no definition of serialVersionUID

Also suppress warning: auxiliary class TargetLanguageBase in ./inherit_target_language.java should not be accessed from outside its own source file

+++ Examples/test-suite/java/cpp11_std_array_runme.java
@@ -69,7 +69,7 @@ public class cpp11_std_array_runme {
 
     // out of range errors
     try {
-      ai.set((int)ai.size(), 0);
+      ai.set(ai.size(), 0);
       throw new RuntimeException("Out of range exception not caught");
     } catch(IndexOutOfBoundsException e) {
     }
